backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
passlib==1.7.4
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0